Add files via upload
Another thing I spent a bit of time on was abstracting text and data from PDFs. Surprisingly (at least to me), there are still not many tools available that do this well, and that includes that new OCR model that Mistral put out. For now, it is straightforward to extract the text, but the formatting of the PDFs poses hurdles, especially in academic papers with multiple different formats depending on the publisher. I have found that pairing DocLayout-YOLO with tesseract can do a decent job, although it still requires effort collating the abstracted texts.